prefer Linux::Inotify2+AnyEvent over AnyEvent::Filesys::Notify
AnyEvent::Filesys::Notify is not optimal, as it gots events for *any*
file change, even for beginning a file write, which is to early.
org-daemon is really only interested in the CLOSE_WRITE event. This
can only be done when using Linux::Inotify2+AnyEvent directly (and
this is only possible on linux systems).